+++ changes.md
@@ -4,6 +4,9 @@
 <summary>Added Property(s)</summary>
 
 - added property `warnings` to type `ProductTailoring`
+- added property `attributes` to type `ProductVariantTailoring`
+- added property `attributes` to type `ProductVariantTailoringDraft`
+- added property `attributes` to type `ProductTailoringAddVariantAction`
 - added property `warnings` to type `Product`
 </details>
 
@@ -11,7 +14,20 @@
 <details>
 <summary>Added Type(s)</summary>
 
+- added type `SearchNotReadyError`
+- added type `GraphQLSearchNotReadyError`
+- added type `ProductTailoringAttribute`
+- added type `ProductTailoringSetAttributeAction`
+- added type `ProductTailoringSetAttributeInAllVariantsAction`
 - added type `ImageProcessingOngoingWarning`
 - added type `WarningObject`
+</details>
+
+
+<details>
+<summary>Added Enum(s)</summary>
+
+- added enum `product-tailoring` to type `MessageSubscriptionResourceTypeId`
+- added enum `product-tailoring` to type `ResourceTypeId`
 </details>